% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_main.R
\name{create_assayv5}
\alias{create_assayv5}
\title{Create Assay V5 Object}
\usage{
create_assayv5(assay_data)
}
\arguments{
\item{assay_data}{A named list of matrices}
}
\value{
Returns a Seurat Assay V5 Object
}
\description{
Given a named list of matrices, create a Seurat Assay V5 Object
}
