% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_feather}
\alias{load_feather}
\title{Load Feather File}
\usage{
load_feather(path)
}
\arguments{
\item{path}{File path}
}
\description{
Read a feather file, determine the objects class based on
its metadata, and return the recreated the object
}
