% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_main.R
\name{load_seurat}
\alias{load_seurat}
\title{Load Seurat Object}
\usage{
load_seurat(dir_path)
}
\arguments{
\item{dir_path}{file path to the saved Seurat object}
}
\value{
A Seurat object
}
\description{
Loads a Seurat object that has been written with save_seurat
}
