% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_main.R
\name{save_seurat}
\alias{save_seurat}
\title{Save Seurat Object}
\usage{
save_seurat(sobj, dir_path = "", compression = "lz4", compression_level = NULL)
}
\arguments{
\item{sobj}{A Seurat object}

\item{dir_path}{A file path to a directory}

\item{compression}{The type of compression to use, default is "zstd"}

\item{compression_level}{If compression "zstd" is specified, the level of compression may also be specified}
}
\value{
Outputs a folder with the file path specified
}
\description{
Save a Seurat object to a directory
}
