% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{write_dataframe}
\alias{write_dataframe}
\title{Write Dataframe}
\usage{
write_dataframe(
  df,
  path,
  name_of_rows = "rownames",
  compression = "lz4",
  compression_level = NULL
)
}
\arguments{
\item{df}{A dataframe}

\item{path}{A file path to write to}

\item{name_of_rows}{The name of the column that the rownames of the dataframe will have}

\item{compression}{The type of compression to use}

\item{compression_level}{The level of compression to use if "zstd" is specified}
}
\description{
Write out a dataframe as a feather file:
'x': the data written out as a table.
}
