% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{write_dense_int}
\alias{write_dense_int}
\title{Write Dense Integer Matrix}
\usage{
write_dense_int(mat, path, compression = "lz4", compression_level = NULL)
}
\arguments{
\item{mat}{A matrix with dimension names}

\item{path}{A file path to write to}

\item{compression}{The type of compression to use}

\item{compression_level}{The level of compression to use if "zstd" is specified}
}
\description{
Write out a dense matrix with integer elements as 3 separate
feather files:
'x': the data written out as a single vector.
'colnames': the names of the columns.
'rownames': the names of the rows.
}
